Moved grid world examples to examples directory
