+++ abmarl/examples/__init__.py
@@ -1,1 +1,3 @@
 from .sim import MultiCorridor
+from .sim import BattleAgent, TeamBattleSim
+from .sim import MazeNavigationAgent, MazeNaviationSim
