+++ abmarl/examples/sim/__init__.py
@@ -1,1 +1,3 @@
 from .multi_corridor import MultiCorridor
+from .team_battle_example import BattleAgent, TeamBattleSim
+from .maze_navigation import MazeNavigationAgent, MazeNaviationSim
