+++ docs/src/tutorials/gridworld/communication.rst
@@ -12,7 +12,8 @@ block these messages to impede consensus. Abmarl's GridWorld Simulation Framewor
 already contains the features for the blocking agents; in this tutorial, we show
 how to create *new* components for the communication feature and connect them with
 the simulation framework. The tutorial can be found in full
-`in our repo <https://github.com/LLNL/Abmarl/blob/main/abmarl/sim/gridworld/examples/comms_blocking.py>`_.
+`in our repo <https://github.com/LLNL/Abmarl/blob/main/abmarl/examples/sim/comms_blocking.py>`_.
+# TODO: Check this ^ reference
 
 .. figure:: /.images/gridworld_tutorial_communications.*
    :width: 75 %
