+++ docs/src/tutorials/gridworld/maze_navigation.rst
@@ -8,7 +8,8 @@ Using the same components as we did in the :ref:`Team Battle tutorial <gridworld
 we can create a Maze Navigation Simulation that contains a single moving agent navigating a maze
 defined by wall agents in the grid. The moving agent's goal is to reach a target
 agent. We will construct the Grid by :ref:`reading a grid file <api_gridworld_sim>`.
-This tutorial can be found in full `in our repo <https://github.com/LLNL/Abmarl/blob/main/abmarl/sim/gridworld/examples/maze_navigation.py>`_.
+This tutorial can be found in full `in our repo <https://github.com/LLNL/Abmarl/blob/main/abmarl/examples/sim/maze_navigation.py>`_.
+# TODO: Check this ^ reference
 
 .. figure:: /.images/gridworld_tutorial_maze_navigation.*
    :width: 75 %
