+++ docs/src/tutorials/gridworld/team_battle.rst
@@ -11,7 +11,8 @@ The goal of each team is to be the last team alive, at which point the simulatio
 Each agent can move around the grid and attack agents from other teams. Each agent
 can observe the grid around its position. We will reward each agent for successful
 kills and penalize them for bad moves. This tutorial can be found in full
-`in our repo <https://github.com/LLNL/Abmarl/blob/main/abmarl/sim/gridworld/examples/team_battle_example.py>`_.
+`in our repo <https://github.com/LLNL/Abmarl/blob/main/abmarl/examples/sim/team_battle_example.py>`_.
+# TODO: Check this ^ reference
 
 .. figure:: /.images/gridworld_tutorial_team_battle.*
    :width: 75 %
